PMBus + HIF (#178)
This work adds a new debugging interface to Hubris: HIF, a simple
bytecode that allows for an in-Hubris agent (task-hiffy) to execute
operations at the direction of Humility. This allows for debugging
interfaces to Hubris tasks: by adding HIF support for a particular
task and operations, it becomes straightforward to implement a
Humility command that executes the operation and interprets the
results. This work reimplements extant debugging interfaces into Jefe
and i2c in terms of HIF and then (less unrelatedly than it might
appear) also implements: PMBus, a small handful of minimal drivers for
PMBus devices (ADM1272, TPS546B24A), and a primordial power loop.

+++ drv/stm32h7-gpio/src/main.rs
@@ -248,13 +248,13 @@ fn main() -> ! {
                     }
                     Op::ReadInput => {
                         let (msg, caller) = msg
-                            .fixed::<u8, u32>()
+                            .fixed::<u8, u16>()
                             .ok_or(ResponseCode::BadArg)?;
                         let port =
                             Port::from_u8(*msg).ok_or(ResponseCode::BadArg)?;
                         let reg = gpio_reg(port);
 
-                        caller.reply(reg.idr.read().bits());
+                        caller.reply(reg.idr.read().bits() as u16);
                     }
                     Op::Toggle => {
                         let (msg, caller) = msg
